Bug #47114: Fix memory leak - __del__
The Python garbage collector (GC) nowadays can free cyclic structure,
but only if that cycle does NOT contain objects implementing the
`__del__` method, as the GC then cannot decide which object to free
first: That method is only called *after* the GC has decided to free the
object and then it is pointless to clean the exception data.

Remove the `__del__` method.

<https://docs.python.org/2/library/gc.html#gc.garbage>