Bug #47114: Fix memory leak - traceback
The Python garbage collector (GC) nowadays can free cyclic structure,
but keeping a Traceback object is still considered a bad idea: It
contains a reference to all stack frames, which reference all local
variables, and thus the traceback itself.

Format the traceback once and replace the reference with None.
The standard traceback library handles this fine.

<https://docs.python.org/2/library/sys.html#sys.exc_info>